pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "open_api_framework"
version = "0.9.2"
description = "A metapackage for registration components, that bundles the dependencies shared between these components and provides generic settings"
authors = [
    {name = "Maykin Media", email = "support@maykinmedia.nl"}
]
readme = "README.rst"
license = {file = "LICENSE"}
keywords = ["metapackage", "dependencies", "settings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "django>=4.2.17,<5.0.0",
    "django-axes>=6.5.1",
    "django-cors-headers>=4.4.0",
    "django-markup>=1.8.1",
    "django-jsonform>=2.22.0",
    "django-log-outgoing-requests>=0.6.1",
    "django-admin-index>=3.1.0",
    "django-redis>=5.4.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-gis>=1.0",
    "django-filter>=24.2",
    "drf-spectacular[sidecar]>=0.27.2",
    "django-csp>=3.8",
    "djangorestframework-inclusions>=1.2.0",
    "commonground-api-common>=1.12.1",
    "mozilla-django-oidc-db>=0.19.0",
    "zgw-consumers>=0.27.0",
    "psycopg2>=2.9.9",
    "bleach>=6.1.0",
    "python-dotenv>=1.0.0",
    "python-decouple>=3.8",
    "requests>=2.32.3",
    "uwsgi>=2.0.23",
    "sentry-sdk>=2.11.0",
    "elastic-apm>=6.22.0",
    "celery>=5.4.0",
    "flower>=2.0.1",
    "maykin-2fa>=1.0.1",
    "django-setup-configuration>=0.6.0",
    "django-sessionprofile>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/maykinmedia/open_api_framework"
Documentation = "http://open_api_framework.readthedocs.io/en/latest/"
"Bug Tracker" = "https://github.com/maykinmedia/open_api_framework/issues"
"Source Code" = "https://github.com/maykinmedia/open_api_framework"

[project.optional-dependencies]
tests = [
    "pytest",
    "pytest-django",
    "tox",
    "isort",
    "black",
    "flake8",
    "factory-boy",
]
coverage = [
    "pytest-cov",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
]
release = [
    "bump-my-version",
    "twine",
]

[tool.setuptools.packages.find]
include = ["open_api_framework*"]
namespaces = false

[tool.isort]
profile = "black"
combine_as_imports = true
known_django = "django"
known_first_party="open_api_framework"
sections=["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pytest.ini_options]
testpaths = ["tests"]
DJANGO_SETTINGS_MODULE = "testapp.settings"

[tool.bumpversion]
current_version = "0.9.2"
files = [
    {filename = "pyproject.toml"},
    {filename = "README.rst"},
    {filename = "docs/conf.py"},
]

[tool.coverage.report]
exclude_also = [
    "if (typing\\.)?TYPE_CHECKING:",
    "@(typing\\.)?overload",
    "class .*\\(.*Protocol.*\\):",
    "@(abc\\.)?abstractmethod",
    "raise NotImplementedError",
    "\\.\\.\\.",
    "pass",
]
